refactor(access-service): extract recall read-surface into access-recall-surface.ts
God-file decomposition continuing the #1526 playbook (verbatim move +
live selfDeps wiring so Object.create(prototype) fakes and instance
stubs keep working). Moves executeRecall, recallXray,
serializeRecallResults, buildRecallResponseFromXraySnapshot, and
fetchRawExcerpts (1,364 LOC, 18 deps) into AccessRecallSurface.
qmdCollectionPathParts/qmdResultPathCandidates come from the canonical
orchestration/qmd-result-resolver copies (access-service keeps its
private duplicates for remaining callers; dedupe tracked separately).

Stacked on refactor/1526-seam27 (needs orchestration/self-deps.ts).

access-service.ts: 9,038 -> 7,849 LOC.
